CHAT-11 Refactor folder structure.
