+++ src/app/app.module.ts
@@ -4,9 +4,8 @@ import {ReactiveFormsModule} from '@angular/forms';
 import {AppRoutingModule} from './app-routing.module';
 import {AppComponent} from './app.component';
 
-import {LoginComponent} from './auth/login/login.component';
-import {SignUpComponent} from './auth/signup/signup.component';
-import {HeaderComponent} from './header/header.component';
+import {LoginComponent} from './core/authentication/login/login.component';
+import {SignUpComponent} from './core/authentication/signup/signup.component';
 import { HttpClientModule } from '@angular/common/http';
 import { GraphQLModule } from './graphql/graphql.module';
 
@@ -14,8 +13,7 @@ import { GraphQLModule } from './graphql/graphql.module';
   declarations: [
     AppComponent,
     SignUpComponent,
-    LoginComponent,
-    HeaderComponent
+    LoginComponent
   ],
   imports: [
     BrowserModule,
