+++ src/app/core/authentication/auth.model.ts
