+++ src/app/core/authentication/signup/signup.component.html
@@ -1,7 +1,7 @@
 <div class="flex h-screen bg-gray-800 text-white bg-primary-black">
   <div class="w-full flex flex-col min-h-screen justify-center items-center lg:w-[36rem] xl:w-[44rem]">
     <div class="p-4 rounded-3xl shadow-md shadow-accent-black">
-      <img src="./assets/images/logo.png" alt="" class="w-auto h-32 object-cover">
+      <img src="../../../../assets/images/logo.png" alt="" class="w-auto h-32 object-cover">
     </div>
     <h1 class="text-6xl font-semibold mt-12">Chatty</h1>
     <h2 class="text-lg italic pt-8">Just connect everywhere</h2>
@@ -68,6 +68,6 @@
     </div>
   </div>
   <div class="flex flex-1 min-h-screen hidden md:block">
-    <img src="./assets/images/signup-page.png" alt="" class="w-full h-full object-cover object-right-bottom">
+    <img src="../../../../assets/images/signup-page.png" alt="" class="w-full h-full object-cover object-right-bottom">
   </div>
 </div>
